+++ MinyanMate/src/main/java/org/minyanmate/minyanmate/database/MinyanSchedulesTable.java
@@ -132,8 +132,12 @@ public class MinyanSchedulesTable {
 	
 	public static void onUpgrade(SQLiteDatabase database, int oldVersion,
 			int newVersion) {
-		database.execSQL("DROP TABLE IF EXISTS " + TABLE_MINYAN_SCHEDULES);
-		onCreate(database);
+        if (oldVersion == 1) {
+            database.execSQL("DROP TABLE IF EXISTS " + TABLE_MINYAN_SCHEDULES);
+            onCreate(database);
+        } else if (oldVersion == 2 && newVersion == 3) {
+            // do nothing!
+        }
 	}
 	
 	private static final SparseArray<String> days = new SparseArray<String>(7);
