+++ MinyanMate/src/main/java/org/minyanmate/minyanmate/database/MinyanSubscriptionsTable.java
@@ -41,7 +41,11 @@ public class MinyanSubscriptionsTable {
 	
 	public static void onUpgrade(SQLiteDatabase database, int oldVersion,
 			int newVersion) {
-		database.execSQL("DROP TABLE IF EXISTS " + TABLE_SUBSCRIPTIONS);
-		onCreate(database);
+        if (oldVersion == 1) {
+            database.execSQL("DROP TABLE IF EXISTS " + TABLE_SUBSCRIPTIONS);
+            onCreate(database);
+        } else if (oldVersion == 2 && newVersion == 3) {
+            // do nothing!
+        }
 	} 
 }
